+++ stream.c
@@ -115,14 +115,14 @@ int rs_buffers_copy(rs_buffers_t *stream, int max_len)
     if ((unsigned) len > stream->avail_in) {
         rs_trace("copy limited to %ld available input bytes",
                  (long) stream->avail_in);
-        len = stream->avail_in;
+        len = (int) stream->avail_in;
     }
 
 
     if ((unsigned) len > stream->avail_out) {
         rs_trace("copy limited to %ld available output bytes",
                  (long) stream->avail_out);
-        len = stream->avail_out;
+        len = (int) stream->avail_out;
     }
 
     if (!len)
